requirements.txt
Flask>=2.3.0
Flask-Session>=0.5.0
pandas>=1.5.0
google-generativeai>=0.3.0
reportlab>=4.0.0
openpyxl>=3.10.0
yfinance>=0.2.0
requests>=2.31.0
cachetools>=5.3.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
SQLAlchemy>=2.0.0

